remove many unnecesary files
