+++ app.json
@@ -1,59 +1,57 @@
 {
-  "expo": {
-    "name": "music",
-    "slug": "music",
-    "version": "1.0.0",
-    "orientation": "portrait",
-    "icon": "./src/assets/images/icon.png",
-    "scheme": "myapp",
-    "userInterfaceStyle": "automatic",
-    "newArchEnabled": false,
-    "assets": [
-      "./src/assets/audio/*"
-    ],
-    "ios": {
-      "supportsTablet": true
-    },
-    "android": {
-      "adaptiveIcon": {
-        "foregroundImage": "./src/assets/images/adaptive-icon.png",
-        "backgroundColor": "#ffffff"
+   "expo": {
+      "name": "music",
+      "slug": "music",
+      "version": "1.0.0",
+      "orientation": "portrait",
+      "icon": "./src/assets/images/icon.png",
+      "scheme": "myapp",
+      "userInterfaceStyle": "automatic",
+      "newArchEnabled": false,
+      "assets": [],
+      "ios": {
+         "supportsTablet": true
       },
-      "package": "com.nicolascervino.music",
-      "permissions": [
-        "android.permission.READ_EXTERNAL_STORAGE",
-        "android.permission.READ_MEDIA_AUDIO"
-      ]
-    },
-    "web": {
-      "bundler": "metro",
-      "output": "static",
-      "favicon": "./src/assets/images/favicon.png"
-    },
-    "plugins": [
-      "expo-router",
-      [
-        "expo-splash-screen",
-        {
-          "image": "./src/assets/images/splash-icon.png",
-          "imageWidth": 200,
-          "resizeMode": "contain",
-          "backgroundColor": "#ffffff"
-        }
+      "android": {
+         "adaptiveIcon": {
+            "foregroundImage": "./src/assets/images/adaptive-icon.png",
+            "backgroundColor": "#ffffff"
+         },
+         "package": "com.nicolascervino.music",
+         "permissions": [
+            "android.permission.READ_EXTERNAL_STORAGE",
+            "android.permission.READ_MEDIA_AUDIO"
+         ]
+      },
+      "web": {
+         "bundler": "metro",
+         "output": "static",
+         "favicon": "./src/assets/images/favicon.png"
+      },
+      "plugins": [
+         "expo-router",
+         [
+            "expo-splash-screen",
+            {
+               "image": "./src/assets/images/splash-icon.png",
+               "imageWidth": 200,
+               "resizeMode": "contain",
+               "backgroundColor": "#ffffff"
+            }
+         ],
+         "expo-sqlite"
       ],
-      "expo-sqlite"
-    ],
-    "experiments": {
-      "typedRoutes": true
-    },
-    "extra": {
-      "router": {
-        "origin": false
+      "experiments": {
+         "typedRoutes": true
       },
-      "eas": {
-        "projectId": "af5409c5-c757-457e-8bee-e91db41fbad1"
-      }
-    },
-    "owner": "nicolascervino"
-  }
+      "extra": {
+         "router": {
+            "origin": false
+         },
+         "eas": {
+            "projectId": "af5409c5-c757-457e-8bee-e91db41fbad1"
+         }
+      },
+      "owner": "nicolascervino"
+   }
 }
