--- src/app/(tabs)/albums.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-export default function Albums() {
-  return <Text>Albums</Text>;
-}
