--- src/app/(tabs)/artists.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-export default function Artists() {
-  return <Text>Artists</Text>;
-}
