--- src/app/(tabs)/index.tsx
@@ -1,16 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function Home() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-        backgroundColor: "#121212",
-      }}
-    >
-      <Text style={{ color: "#FFFFFF" }}>Edit app/index.tsx to edit this screen.</Text>
-    </View>
-  );
-}
