--- src/app/(tabs)/playlists.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-
-export default function Playlists() {
-  return (
-    <View>
-      <Text style={{ color: "#FFFFFF" }}>Playlists</Text>
-    </View>
-  )
-};
